Azure Function App/requirements.txt
azure-functions
azure-eventhub
azure-identity
azure-keyvault-secrets